Fix DecodedVector's wrap() and dictionaryWrapping() API
This fix addresses multiple inconsistencies in implementation of
wrap() and dictionaryWrapping() which has caused multiple bugs
over the past year that surfaced due to its use in FieldReference
and Cast Expressions. This primarily occurred due to the difference
between how DecodedVector combines indices and nulls and the
changing expectations from the aforementioned APIs. The class
comment for DecodedVector is also expanded to explain its inner
workings and clarify its expected behavior.

Test Plan:
Test cases are added for wrap that cover all possible cases
that can result in different outcomes.